requirements.txt
selenium
requests
python-dotenv
selenium-wire
pydantic